Update Migrations to work with MySQL database.

+++ database/migrations/2025_03_19_154553_create_shelves_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
         });
 
         Schema::table('containers', function (Blueprint $table) {
-            $table->integer('shelf_id');
+            $table->unsignedBigInteger('shelf_id');
             $table->foreign('shelf_id')->references('id')->on('shelves')->onDelete('cascade');
         });
